+++ src/Example/Program.cs
@@ -2,6 +2,7 @@
 using System.Globalization;
 using System.Linq;
 using Isop;
+using Isop.Api;
 
 namespace Example
 {
