+++ src/Isop/CommandLine/ArgumentInvoker.cs
@@ -12,12 +12,12 @@ namespace Isop.CommandLine
 {
     public class ArgumentInvoker
     {
-        private IServiceProvider _serviceProvider;
+        private readonly IServiceProvider _serviceProvider;
         private Recognizes _recognizes;
         private ILookup<string, ArgumentAction> _recognizesMap;
-        private Recognizes recognizes => _recognizes ??= _serviceProvider.GetRequiredService<Recognizes>();
-        private ILookup<string,ArgumentAction> recognizesMap => _recognizesMap 
-            ??= recognizes.Properties.Where(p=>p.Action!=null).ToLookup(p=>p.Name, p=>p.Action);
+        private Recognizes Recognizes => _recognizes ??= _serviceProvider.GetRequiredService<Recognizes>();
+        private ILookup<string,ArgumentAction> RecognizesMap => _recognizesMap 
+            ??= Recognizes.Properties.Where(p=>p.Action!=null).ToLookup(p=>p.Name, p=>p.Action);
 
         public ArgumentInvoker(IServiceProvider serviceProvider)
         {
@@ -26,14 +26,20 @@ namespace Isop.CommandLine
 
         public async Task<IEnumerable> Invoke(ParsedArguments parsedArguments)
         {
+            IEnumerable<Task<object>> Choose(RecognizedArgument arg)
+            {
+                return RecognizesMap.Contains(arg.Argument.Name) 
+                    ? RecognizesMap[arg.Argument.Name].Select(action => action(arg.Value)) 
+                    : Enumerable.Empty<Task<object>>();
+            }
+
             using (var scope = _serviceProvider.CreateScope())
             {
                 return await parsedArguments.Map<Task<IEnumerable>>(
                     @default: async args =>
                     {
-                        var enumerable = args.RecognizedArguments.SelectMany(arg => recognizesMap.Contains(arg.Argument.Name) 
-                            ? recognizesMap[arg.Argument.Name].Select(action => action(arg.Value)) 
-                            : Enumerable.Empty<Task<object>>());
+                        var enumerable = args.RecognizedArguments
+                            .SelectMany(Choose);
                         return await Task.WhenAll(enumerable);
                     },
                     merged: async merged =>
