+++ src/Isop/Configuration.cs
@@ -10,7 +10,7 @@ namespace Isop
         /// <summary>
         /// The culture info for the console application
         /// </summary>
-        public CultureInfo CultureInfo { get; set; } = CultureInfo.InvariantCulture;
+        public CultureInfo CultureInfo { get; set; }
         /// <summary>
         /// If you want to disable infer parameter
         /// </summary>
