+++ src/Tests/DisposeTests.cs
@@ -1,10 +1,11 @@
-using System;
 using System.Globalization;
 using System.IO;
-using Isop.Tests.FakeControllers;
+using Isop;
+using Isop.Api;
 using NUnit.Framework;
+using Tests.FakeControllers;
 
-namespace Isop.Tests
+namespace Tests
 {
 
     [TestFixture]
