+++ src/Tests/FakeControllers/DisposeController.cs
@@ -1,5 +1,6 @@
 ﻿using System;
-namespace Isop.Tests.FakeControllers
+
+namespace Tests.FakeControllers
 {
     public class DisposeController:IDisposable
     {
