+++ src/Tests/FakeControllers/EnumerableController.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Isop.Tests.FakeControllers
+namespace Tests.FakeControllers
 {
     class EnumerableController
     {
