+++ src/Tests/FakeControllers/ObjectController.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Isop.Tests
+namespace Tests.FakeControllers
 {
     public class ObjectController
     {
