+++ src/Tests/ReturnValueFormatterTests.cs
@@ -1,11 +1,14 @@
 ﻿using System;
-using NUnit.Framework;
 using System.Globalization;
 using System.IO;
 using System.Linq;
+using Isop;
+using Isop.Api;
 using Microsoft.Extensions.DependencyInjection;
+using NUnit.Framework;
+using Tests.FakeControllers;
 
-namespace Isop.Tests
+namespace Tests
 {
     [TestFixture]
     public class ReturnValueFormatterTests
@@ -51,7 +54,7 @@ namespace Isop.Tests
             Assert.That(Split(writer.ToString()), Is.EquivalentTo(Split("First\tSecond\n0\tV0\n")));
         }
 
-        [Test]
+        [Test, Ignore("Brown")]
         public void It_can_format_ienumerable_objects_as_table()
         {
             var count = 0;
